requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
Flask==2.0.3
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-datastore==1.15.3
google-cloud-ndb==1.11.1
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymemcache==3.5.1
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
